requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.2.0
tqdm>=4.65.0
typing-extensions>=4.0.0

# Data processing
pandera>=0.16.0
pydantic>=2.0.0
pydantic-core>=2.0.0
python-dotenv>=1.0.0
pytz>=2023.3
PyYAML>=6.0
email-validator>=2.0.0  # Required by pydantic for email validation
typeguard>=4.0.0
typing-inspect>=0.8.0

# Database
pymongo>=4.3.0
sqlalchemy>=2.0.0
sqlmodel>=0.0.8

# Financial data
yfinance>=0.2.28
yahooquery>=2.3.0
yahoo-fin>=0.8.9
alpha_vantage>=2.3.1
pandas-ta>=0.3.14b
nsepython>=2.9.0  # Indian stock market data

# Machine learning
tensorflow>=2.12.0
keras>=2.12.0
xgboost>=1.7.5
lightgbm>=3.3.5
optuna>=3.3.0
mlflow>=2.6.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
imbalanced-learn>=0.10.0

# Web and API
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
starlette>=0.27.0
sse-starlette>=1.0.0
h11>=0.14.0
httpcore>=0.17.0
httptools>=0.5.0
watchfiles>=0.18.0
websockets>=10.0

# Authentication and security
PyJWT>=2.0.0
bcrypt>=4.0.0

# Intelligence
langchain>=0.0.310
chromadb>=0.4.15
openai>=0.28.0
transformers>=4.30.0
tokenizers>=0.13.0
huggingface-hub>=0.15.0
faiss-cpu>=1.7.0

# Web scraping and data collection
requests>=2.30.0
requests-html>=0.10.0
beautifulsoup4>=4.11.0
bs4>=0.0.1
lxml>=4.9.0

# Visualization
plotly>=5.15.0
dash>=2.13.0
dash-bootstrap-components>=1.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.82.0

# Development
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.4.0
pre-commit>=3.3.3
